+++ examples/web-components-wrapper/src/main.rs
@@ -1,5 +1,7 @@
 use futures_signals::signal::{Mutable, SignalExt};
-use silkenweb::{elements::html::*, log_panics, prelude::*, value::Sig};
+use silkenweb::{
+    elements::html::*, log_panics, node::element::TextParentElement, prelude::*, value::Sig,
+};
 use silkenweb_example_web_components_wrapper::{ui5_button, ui5_button_badge, ButtonDesign};
 
 fn main() {